pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spoon-bot"
version = "0.1.0"
description = "Local-first AI agent with native OS tools, powered by spoon-core"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "XSpoon Team", email = "team@xspoon.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
keywords = ["ai", "agent", "llm", "local", "assistant"]

dependencies = [
    # Required: spoon-core SDK for LLM, MCP, and skills (v0.4.0+)
    "spoon-ai-sdk>=0.4.0",
    # CLI and utilities
    "typer>=0.9.0",
    "rich>=13.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",  # For SKILL.md parsing
]

[project.optional-dependencies]
toolkit = [
    "spoon-toolkits>=0.1.0",  # 40+ tools
]
memory = [
    "mem0ai>=0.0.1",  # Semantic memory
]
web3 = [
    "web3>=6.0.0",
]
document = [
    "pymupdf>=1.24.0",      # PDF parsing (text, tables, images)
    "openpyxl>=3.1.0",      # Excel (.xlsx) generation
    "Pillow>=10.0.0",       # Image processing
    "python-pptx>=0.6.21",  # PowerPoint generation
]
gateway = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "PyJWT>=2.8.0",
    "websockets>=12.0",
]
all = [
    "spoon-bot[core,toolkit,memory,web3,document,gateway]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
spoon-bot = "spoon_bot.cli:app"

[project.urls]
Homepage = "https://github.com/XSpoonAi/spoon-bot"
Repository = "https://github.com/XSpoonAi/spoon-bot"
Documentation = "https://github.com/XSpoonAi/spoon-bot#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["spoon_bot*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
